Decoupled Sticky Component
I created ScrollEventContainer to send window scroll data to a generic
child element.

I created StickyEventComponent that makes all child elements sticky
based on entered attributes. Works perfectly as a child of
ScrollEventContainer.

I also decoupled NavigationListComponent so that it can be reuesd
generically with and array of Link objects.