FINALLY got the Grid to work!
Grid works to my satisfaction. I was hoping to automate more of the
media query behavior but it was way too complicated and a waste of
memory.

Alas, the Grid now has a breakPoints attribute that accepts an array
with up to 5 ints that signify the window width that brings the Grid to
the next breakLevel.

The Cols then have an attribute called breaks that accepts an array
signifying the width of that column as a percentage when the Grid hits a
breakLevel. A Col can remain the same width by repeating the number in
the breaks array.