Fixed the ResponsiveContainer and Grid
The async call to the windowState redux store for the window width was
causing the Grid and ResponsiveContainer to break when the window was
maximized/unmaximized. I now have them getting window.innerwidth
directly.

The windowState store is still necessary, however, in order to have the
Grid and ResponsiveContainer's componentWillUpdate code to fire at the
right time and to have only a single event listener for resizing in
app.js.

I assume the same is true for scrollState, but scrollState can be
slightly asynchronous without causing issues (at least for now).

I also fixed the Col logic to assign the highest breakLevel width when
an individual Col has fewer breaks than the Grid assigns.  It is also
possible for a Col to have no assigned breaks and stay the same size at
all breakLevels.

The Col code is a bit sloppy but I am sick of dealing with these issues
and it works, so screw it for now!