Face Palm Moment
Yep, still not scroll loading.

I never really got 'margin: 0 auto' to work before, so I did not even
try it. But of course it works and it is much better than the elaborate
javascript I was using in ResponsiveContainer.

But now I am using this simple css solution instead. I have also added
the option to override the margin with the rcMargin prop and the option
to override the dirplay with rcDisplay to inline-block, allowing
ResponsiveContainers to be placed next to each other in the DOM.

All of the props for ResponsiveContainer are optional: rcWidth (%),
rcMaxWidth (px), rcMinWidth (px), rcMargin (%), and rcDisplay (string
display type; I don't recommend using anything but inline-block as block
is defaulted and inline breaks the ResponsiveContainer) .