InventoryContainer
I started to create InventoryContainer which displays the list of
inventoryState.items names and allows removal of items.

I also updated the Item class to include the Item dimensions, display
and position styles, and an acquired bool that indicates whether or not
the item is already in the inventoryState.items array (which was
previously an object).

The ItemContainer react element will check to see if the Item is
acquired and display a placeholder div if it is. This works across
routes to allow the inventory to be used anywhere in the site!