Changed TimerContainer to EventExectutor
And it may be the coolest thing I have made yet!!!!!!

EventExecutor executes custom functions at specific intervals on its
child element.

It takes the following props:
moments(required): An array of whole numbers that signify the moment in
time that each supplied event function will execute.

events(required): An array of functions that can be custom written to do
pretty much anything to the child of EventExecutor. Should be the same
array length as moments.

loop(defaults to false): Boolean that sets whether or not the
EventExecutor should loop the events.

increment(defaults to 1000): Length of time that the counter is counting
by in milliseconds.

I am pretty much a genius now.