Not Landing Page....
But I did do awesome stuff with EventExecutor (now called
DeferredEventExecutor).

DeferredEventExecutor now takes an optional prop fireCondition(string)
which is the name of an event to search for in the firedEvents array in
the interactiveState store. The events will not fire until the
fireCondition is in the array.

DropZone now has optional prop eventTrigger which defaults to false. If
set to true, the DropZone will add the name of the dropzone plus 'Event'
to the firedEvents array in the interactableState.