CSSAnimationQueuer
So I created a new badass React component called CSSAnimationQueuer.

It wraps its children inside of a defined number of divs and assigns a
classname to the parent div of the div group.

Required props:

queueCount(number):
Specifies the number of divs to nest inside of each other. Should be the
same amount of different animations intended to be executed on the
wrapped children.

animationClass(CSS class string):
The intended className for the parent div of the group. The
corresponding CSS class should be structured so that every separately
keyframed animation is assigned to a different div. Chronology is
determined by the delay set on the animation property, not the div
order. Rotation animations should be nested the deepest as they alter
the translation reference.

CSSAnimationQueuer can also add firedEvents to the Redux
interactableState firedEvents array as well as listen for them if
necessary. This allows CSSAnimationQueuer to both fire events and be
fired by events. Both of these capabilities are optional and mutually
exclusive.

To make an event trigger the animation, use the props:

triggeredByEvent(string):
Specifies the name of the firedEvent to look for in the firedEvents
array in the interactableState store.

initialStyle(CSS style string):
Assigns the initial style of the CSSAnimationQueuer's placeholder child.
Create an initial style CSS class (which can be extended into the
animation class to reduce redundancy and potential mistakes) and assign
it to this prop.

If the CSSAnimationQueuer is to trigger an event, use the props:
eventToTrigger(string):
Name of the event to add to the firedEvents array.

triggerDelay(number):
The length of time (in milliseconds) to delay triggering the
eventToTrigger. It can be the length of the CSS event itself, or any
other number.