3D Stuff
I created a React component ThreeDSpace that takes the prop perspective
which sets the 3d perspective css. It is a flex-box that can contain 3D
objects and be additionally styled as needed.

I also created ThreeDObjectGroup that can parent groups of 3D objects to
allow transformations and animations that affect the whole group. To be
contained inside of a ThreeDSpace.

I also updated Cube to be decoupled from its own world (now the
ThreeDSpace component) and should be placed inside of a ThreeDSpace.

Both Cube and ThreeDObjectGroup can be animated using their public
updateTransform() methods as long as they are inside of an imported
component and have ref tags on them.

I can use DeferredEventExecutor to fire updateTransform() on Cube and
ThreeDObjectGroup with arrays of transform data that allow multiple
transforms to be fired at the same time on the same div. The transforms
also retain all unchanged information so only that which needs to change
can be updated.

This is a terrible and complex description, I know, but these components
are powerful and will allow fairly robust 3D css animations that can be
chained together and triggered by Redux store interactableState
firedEvents!