FIXED INVENTORY!
What a pain! Dragging inventory items is now nice and clean. While
dragging an item, the item has pointerEvents set to none and the
document has a click event listener added that returns the item to the
inventory. This allows items on the page to still be clicked and to fire
a relevent event if the user is currently dragging the correct item.